+++ src/Chunkyard/Program.cs
@@ -38,7 +38,7 @@ public static class Program
         Parser.Default.ParseArguments(args, optionTypes)
             .WithParsed<RestoreOptions>(Commands.RestoreSnapshot)
             .WithParsed<MirrorOptions>(Commands.MirrorSnapshot)
-            .WithParsed<CreateOptions>(Commands.CreateSnapshot)
+            .WithParsed<StoreOptions>(Commands.StoreSnapshot)
             .WithParsed<CheckOptions>(Commands.CheckSnapshot)
             .WithParsed<ShowOptions>(Commands.ShowSnapshot)
             .WithParsed<RemoveOptions>(Commands.RemoveSnapshot)